// actions/dataImport/entryData/types.ts
export enum DataImportTypes {
  SET_ENTRY_DATA_LIST = "DATA_IMPORT_SET_ENTRY_DATA_LIST",
  SET_PARSING = "DATA_IMPORT_SET_PARSING",
  SET_MAPPABLE_DATA = "DATA_IMPORT_SET_MAPPABLE_DATA",
  SET_NUMBER_DATA = "DATA_IMPORT_SET_NUMBER_DATA",
  SET_STRING_DATA = "DATA_IMPORT_SET_STRING_DATA",
  TAG_CLICKED = "DATA_IMPORT_TAG_CLICKED",
  SET_COMPUTE_METHOD = "DATA_IMPORT_SET_COMPUTE_METHOD",
  SET_EMISSION_FACTOR = "DATA_IMPORT_SET_EMISSION_FACTOR",
  SAVE_DATA_REQUESTED = "DATA_IMPORT_SAVE_DATA_REQUESTED",
  DATA_SAVED = "DATA_IMPORT_DATA_SAVED",
  DATA_SAVING_FAILED = "DATA_IMPORT_DATA_SAVING_FAILED",
  ERROR_ACKNOLEDGED = "DATA_IMPORT_ERROR_ACKNOLEDGED",
  DELETE_ENTRY_DATA = "DATA_IMPORT_DELETE_ENTRY_DATA",
  RESET = "DATA_IMPORT_RESET",
  SET_MULTI_ENTRY_TAGS = "DATA_IMPORT_SET_MULTI_ENTRY_TAGS",
  SET_MULTI_ENTRY_INPUT_UNITS = "DATA_IMPORT_SET_MULTI_ENTRY_INPUT_UNITS",
}